storage, backup: extract pull-mode functions to pull_mode.go
Move all pull-mode related functions and constants out of backup.go
into a dedicated file.
